# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cobranca-API — Offline (engine pyCobranca)
  description: |
    Superfície **offline** da plataforma **Cobranca-API** (serviço único **100% Python**):
    geração de boletos bancários brasileiros, arquivos CNAB e parsing de extratos OFX —
    servida **nativamente** pela engine [pyCobrança](https://github.com/Maxwbh/pyCobranca)
    dentro do próprio processo FastAPI (sem sidecar, sem proxy).

    > **Nota de versão (v2.0.0):** os contratos desta superfície (paths, campos,
    > headers `X-*`) permanecem compatíveis com a v1.x. A única mudança de escopo é
    > o formato de saída: a engine emite **PDF** — `jpg`/`png`/`tif` respondem
    > **400**.

    ## Offline (`/api/*`) × Gateway (raiz) — quando usar cada um

    | | **Offline (esta doc)** | **Gateway REST** ([Swagger `/docs`](/docs)) |
    |---|---|---|
    | O que faz | GERA o documento (boleto/CNAB) localmente | REGISTRA a cobrança na API online do banco (C6, Sicoob, Inter, Itaú) |
    | Boleto | Não registrado (o banco só conhece no pagamento) | **Registrado** (CIP, em tempo real) + Pix, Bolepix, webhooks |
    | Autenticação | **Nenhuma** (por design — não toca em banco nem segredo) | Token `bapi_` (`POST /credenciais`) |
    | Conciliação | Retorno CNAB / OFX (arquivos) | Webhooks + consultas online |

    > **Autenticação:** esta superfície não tem `securitySchemes` de propósito.
    > O mecanismo `bapi_` pertence ao gateway e está documentado no
    > [Swagger do gateway](/docs).

...

...

...

...

...
servers:
...
paths:
...
  /api/retorno:
    post:
      tags:
        - Retorno
      summary: Processar arquivo de retorno
      description: Faz parsing de arquivo de retorno CNAB 240 ou 400
      operationId: processRetorno
      parameters:
        - name: bank
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/BankCode'
        - name: type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CnabType'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: string
                  format: binary
                  description: Arquivo de retorno CNAB
      responses:
        '200':
          description: Retorno processado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoResponse'
        '400':
          description: Erro de processamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: |
            Arquivo acima de `UPLOAD_MAX_BYTES` (default **10 MB**). O teto é
            conferido ANTES de carregar o conteúdo — sem ele, um envio grande
            derrubava o processo por memória antes de qualquer validação.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  campo:
                    type: string
                  recebidos:
                    type: integer
                  hint:
                    type: string
  /api/ofx/parse:
    post:
      tags:
        - OFX
      summary: Parsear arquivo OFX (extrato bancário)
      description: |
        Parseia arquivo OFX e retorna JSON estruturado com:
        - Dados do banco (ORG, FID)
        - Dados da conta (agência, número, tipo)
        - Período do extrato
        - Saldo
        - Lista de transações classificadas (crédito/débito)
        - Extração automática de `nosso_numero` do campo memo por banco
        - Resumo com totais de créditos e débitos

        **Recursos:**
        - Suporta OFX v1 (SGML) e v2 (XML)
        - Conversão automática de encoding Latin-1 → UTF-8
        - Extração por regex para Sicoob, Itaú, Banco do Brasil, Bradesco e Caixa
        - Fallback genérico para outros bancos

        **Filtros opcionais:**
        - `somente_creditos=true`: retorna apenas transações de crédito
      operationId: parseOfx
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Arquivo OFX (extrato bancário)
                somente_creditos:
                  type: string
                  enum: ['true', 'false']
                  default: 'false'
                  description: Se 'true', retorna apenas transações de crédito
      responses:
        '201':
          description: Arquivo OFX parseado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfxResponse'
              example:
                banco: { org: ITAU, fid: '341' }
                conta: { agencia: '341', numero: 56789-0, tipo: CHECKING }
                periodo: { inicio: '2025-02-05', fim: '2025-02-10' }
                saldo: { valor: 5230.75, data: '2025-02-28' }
                transacoes:
                  - tipo: credito
                    valor: 150.0
                    data: '2025-02-05'
                    memo: 'PAGTO BOLETO 01234567000000123'
                    nosso_numero: '000000123'
                resumo: { total_creditos: 150.0, total_debitos: 0.0, quantidade: 1 }
        '400':
          description: Arquivo OFX inválido ou ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfxError'
        '413':
          description: |
            Arquivo acima de `UPLOAD_MAX_BYTES` (default **10 MB**). O teto é
            conferido ANTES de carregar o conteúdo — sem ele, um envio grande
            derrubava o processo por memória antes de qualquer validação.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  campo:
                    type: string
                  recebidos:
                    type: integer
                  hint:
                    type: string
...
components:
...
  schemas:
    BankCode:
      type: string
      enum:
        - banco_brasil
        - itau
        - bradesco
        - caixa
        - santander
        - sicoob
        - sicredi
        - banrisul
        - banestes
        - banco_nordeste
        - banco_brasilia
        - banco_c6
        - unicred
        - credisis
        - safra
        - citibank
        - hsbc
        - ailos
      description: Código do banco suportado (18 bancos brasileiros)
...
    CnabType:
      type: string
      enum:
        - cnab240
        - cnab400
      description: Tipo de arquivo CNAB
...
    RetornoResponse:
      type: array
      items:
        $ref: '#/components/schemas/RetornoItem'
    RetornoItem:
      type: object
      description: |
        **Campo sem valor no arquivo não vem na resposta** — a chave é omitida,
        não vem `null`. `numero_documento` e `data_credito` são os que mais
        faltam: em retorno de liquidação simples, nenhum dos dois costuma vir.

        As **datas saem como o banco as escreveu**, sem conversão: `DDMMAA` no
        CNAB 400 (`"150726"` = 15/07/2026) e `DDMMAAAA` no CNAB 240. Não é
        ISO — converta antes de comparar com data do seu sistema.
      properties:
        nosso_numero:
          type: string
          description: Nosso número do título
        numero_documento:
          type: string
          description: Seu número, quando o arquivo o traz.
        data_credito:
          type: string
          description: Data do crédito, no formato do layout (`DDMMAA` / `DDMMAAAA`).
          example: '200726'
        data_ocorrencia:
          type: string
          description: Data da ocorrência, no formato do layout.
          example: '150726'
        valor_titulo:
          type: number
          format: float
        valor_pago:
          type: number
          format: float
        valor_tarifa:
          type: number
          format: float
        codigo_ocorrencia:
          type: string
          description: Código da ocorrência bancária
        motivo_ocorrencia:
          type: string
          description: Descrição da ocorrência
    OfxResponse:
      type: object
      properties:
        banco:
          type: object
          description: Identificação do banco do extrato
          properties:
            org:
              type: string
              description: Organização (ORG do OFX)
              example: SICOOB
            fid:
              type: string
              description: ID do banco (FID do OFX)
              example: '756'
        conta:
          type: object
          description: Dados da conta bancária
          properties:
            agencia:
              type: string
              description: Agência (BANKID do OFX)
              example: '3044'
            numero:
              type: string
              description: Número da conta
              example: '12345-6'
            tipo:
              type: string
              description: Tipo da conta (CHECKING, SAVINGS, etc)
              example: CHECKING
        periodo:
          type: object
          description: Período coberto pelo extrato
          properties:
            inicio:
              type: string
              format: date
              example: '2025-01-01'
            fim:
              type: string
              format: date
              example: '2025-01-31'
        saldo:
          type: object
          properties:
            valor:
              type: number
              format: float
              example: 15420.50
            data:
              type: string
              format: date
              example: '2025-01-31'
        transacoes:
          type: array
          items:
            $ref: '#/components/schemas/OfxTransacao'
        # `total_creditos`/`total_debitos` são SOMAS EM REAIS, não contagens —
        # esta página já disse `integer` com exemplo `28`, e quem conciliava
        # lia R$ 3.750,00 como "28 transações". `soma_creditos`,
        # `soma_debitos` e `total_transacoes` nunca existiram na resposta.
        resumo:
          type: object
          required: [total_creditos, total_debitos, quantidade]
          properties:
            total_creditos:
              type: number
              format: float
              description: Soma dos créditos, em reais.
              example: 3750.00
            total_debitos:
              type: number
              format: float
              description: |
                Soma dos débitos, em reais — sempre positiva. Com
                `somente_creditos=true` vale `0`.
              example: 530.50
            quantidade:
              type: integer
              description: Transações no extrato, depois do filtro.
              example: 4
    OfxTransacao:
      type: object
      description: |
        Seis campos, sempre estes. A página já prometeu `name`, `checknum` e
        `refnum`, que a resposta não traz, e chamava o identificador de
        `fitid` — ele vem em **`id`**, e é por ele que se deduplica ao
        reprocessar o mesmo período.
      required: [tipo, valor, data, memo, id, nosso_numero]
      properties:
        id:
          type: string
          nullable: true
          description: |
            FITID do arquivo — identificador único da transação no extrato.
            `null` quando o arquivo não o traz.
          example: '202501150001'
        tipo:
          type: string
          enum: [credito, debito]
          description: |
            Direção do lançamento, **normalizada** — não é o TRNTYPE cru do
            arquivo. Sai do TRNTYPE e não do sinal, porque a engine devolve
            `valor` sempre positivo.
          example: credito
        data:
          type: string
          format: date
          example: '2025-01-15'
        valor:
          type: number
          format: float
          description: Valor absoluto (sempre positivo — a direção está em `tipo`)
          example: 1250.00
        memo:
          type: string
          description: Descrição da transação
          example: COBRANCA SICOOB 0000012345
        nosso_numero:
          type: string
          nullable: true
          description: Nosso número extraído do memo com a regra do banco emissor
          example: '0000012345'
    OfxError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: |
            Mensagem de erro. `error` é a chave canônica em toda a superfície;
            esta rota respondia só `erro` (em português), que segue presente
            como alias para não quebrar quem já lê por ele.
          example: Arquivo OFX inválido
        erro:
          type: string
          description: Alias de `error` (compatibilidade).
          example: Arquivo OFX inválido
        validation_errors:
          type: array
          description: Detalhe do problema como lista (contrato PyCobrança).
          items:
            type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: Dados inválidos
        validation_errors:
          type: array
          description: Lista plana — um item por campo violado (contrato PyCobrança ≥ 1.0.1).
          items:
            type: string
        details:
          type: array
          description: |
            Só em `/api/retorno`, que usa este nome no lugar de
            `validation_errors` — um item por problema do arquivo.
          items:
            type: string
        hint:
          type: string
          description: O que fazer, quando a rota tem o que sugerir.
          example: Verifique se todos os campos obrigatórios estão preenchidos